Add developer mode.
Developer Mode can be enabled by tapping the title in the app bar on
the start screen five times. Once enabled, a new "Developer Settings"
screen is available in Settings. Right now it allows toggling L2CAP on
and off.

Also, in Developer Mode it's possible to double-tap the portrait or
the error icon to get to a screen with detailed information about the
response including how long it took to establish a connection to the
remote device.

Test: Manually tested
Signed-off-by: David Zeuthen <zeuthen@google.com>